+++ pkg/server/world/object.go
@@ -20,5 +20,5 @@ func (o *Object) Equals(o1 interface{}) bool {
 
 //NewObject Returns a reference to a new instance of a game object.
 func NewObject(id, direction, x, y int, boundary bool) *Object {
-	return &Object{id, direction, boundary, Entity{Location{X: x, Y: y}, -1}}
+	return &Object{id, direction, boundary, Entity{Location{X: uint32(x), Y: uint32(y)}, -1}}
 }
